% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/along.R
\name{along}
\alias{along}
\alias{list_along}
\title{Helper to create vectors with matching length.}
\usage{
list_along(x)
}
\arguments{
\item{x}{A vector.}

\item{y}{Values to repeat.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
These functions take the idea of \code{\link[=seq_along]{seq_along()}} and generalise
it to creating lists (\code{list_along}) and repeating values
(\code{rep_along}).
}
\examples{
x <- 1:5
rep_along(x, 1:2)
rep_along(x, 1)
list_along(x)
}
\keyword{internal}
