% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Chart the application of a function along coordinates}
\usage{
chart(..f, .route, ...)
}
\arguments{
\item{..f}{A function that takes a list of length equal to the
cardinality of the coordinates in \code{.route}.}

\item{.route}{A list of indices used to subset \code{.x}; a
character or numeric vector; or a function that takes .x and returns
a sequence of indices, like \code{seq_window()}, \code{seq_pairs()}
or \code{seq_along()}.}

\item{...}{Further arguments passed on to the charted
function. They are computed only once when \code{chart()} is
called.}
}
\value{
A function that takes a list \code{.x} as first argument.
}
\description{
\code{chart()} modifies a function that takes a list to make it act
on a given route of coordinates. A charted function applies itself
on elements of a list according to a sequence of indices defined by
\code{.route}.
}
\examples{
# We can use chart() to create a function that checks that
# all consecutive pairs of elements of .x are identical()
check_identical <- chart(identical, .route = seq_window)
mtcars[c(1, 1, 2, 3, 3)] \%>\% check_identical()
}
\seealso{
\code{\link{map_along}()}, \link{sequences}
}

