% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{decompose}
\title{Compose multiple functions}
\usage{
compose(...)

decompose(x)
}
\arguments{
\item{...}{n functions to apply in order from right to left.}

\item{x}{Composite function, i.e., a composition of functions created by
\code{compose()}.}
}
\value{
A function
}
\description{
Compose multiple functions
}
\examples{
not_null <- compose(`!`, is.null)
not_null(4)
not_null(NULL)

add1 <- function(x) x + 1
compose(add1, add1)(8)
}
