% Generated by roxygen2 (4.0.2): do not edit by hand
\name{contains}
\alias{contains}
\title{Does a list contain an object?}
\usage{
contains(.x, .y)
}
\arguments{
\item{.x}{A list or vector.}

\item{.y}{Object to test for}
}
\description{
Does a list contain an object?
}
\examples{
x <- list(1:10, 5, 9.9)
x \%>\% contains(1:10)
x \%>\% contains(3)
}

