% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create an environment and export variables to it.}
\usage{
create_env(.vars, hash = TRUE, parent = globalenv(), size = 29L)
}
\arguments{
\item{.vars}{A named list of variables to export to the environment. The names will
be the new names of the objects in the new environment.}

\item{hash}{a logical, if \code{TRUE} the environment will use a hash table.}

\item{parent}{an environment to be used as the enclosure of the
    environment created.}

\item{size}{an integer specifying the initial size for a hashed
    environment.  An internal default value will be used if
    \code{size} is \code{NA} or zero.  This argument is ignored if
    \code{hash} is \code{FALSE}.}
}
\value{
The new environment containing the exported variables.
}
\description{
Creates an environment and exports a list of variables to it.
The default parent is .GlobalEnv.
}
\details{
When calling several functions outside of .GlobalEnv, the entire calling
environment is saved with the object. This can lead to unnecessary information being
retained in memory or saved to disk. Calling a function in an environment
with only the objects required prevents this from happening.
}
\examples{
env <- create_env(list(iris_new_env = iris))
}

