% Generated by roxygen2 (4.1.0): do not edit by hand
\name{cross}
\alias{cross}
\title{Produce all combinations of list elements}
\usage{
cross(.x)
}
\arguments{
\item{.x}{A list.}
}
\description{
\code{cross} returns the product set of the elements of
\code{.x}. It is similar to \code{expand.grid} but it takes and
returns a list.
}
\examples{
# We build all combinations of names, greetings and separators from our
# list of data and pass each one to paste()
data <- list(
  id = c("John", "Jane"),
  greeting = c("Hello.", "Bonjour."),
  sep = c("! ", "... ")
)
args <- cross(data)

out <- vector("list", length(args[[1]]))
for (i in seq_along(out))
  out[[i]] <- map(args, i) \%>\% splat(paste)()
out
}
\seealso{
expand.grid
}

