% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{detect}
\alias{detect}
\alias{position}
\title{Find the value or position of the first match.}
\usage{
detect(.x, .f, ...)

position(.x, .f, ...)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A predicate function, i.e. a function that returns either
\code{TRUE} or \code{FALSE}}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\value{
\code{detect} the value of the first item that matches the
 predicate; \code{position} the position of the matching item.
}
\description{
Find the value or position of the first match.
}
\examples{
is_even <- function(x) x \%\% 2 == 0

1:6 \%>\% detect(is_even)
1:6 \%>\% position(is_even)
}

