% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{done_box}
\alias{done_box}
\alias{is_done_box}
\title{Box a final value for early termination}
\usage{
done_box(x)

is_done_box(x)
}
\arguments{
\item{x}{For \code{done_box()}, a value to box. For \code{is_done_box()}, a
value to test.}
}
\description{
A value boxed with \code{done_box()} signals to its caller that it
should stop iterating. Use it to shortcircuit a loop.

Currently, \code{\link[=reduce]{reduce()}} and \code{\link[=reduce2]{reduce2()}} handle done boxes.
}
\examples{
done_box(3)

x <- done_box(3)
is_done_box(x)
}
