% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{empty}
\alias{empty}
\title{Is a vector/list empty?}
\usage{
empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is a vector/list empty?
}
\examples{
empty(NULL)
empty(list())
empty(list(NULL))
}

