% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mapper.R
\name{extract}
\alias{extract}
\title{Extract an element from a vector or environment}
\usage{
extract(x, index, default = NULL)
}
\arguments{
\item{x}{A vector or environment}

\item{index}{A list indexing into the object}

\item{default}{Value to use if target is empty or absent.}
}
\description{
Extract an element from a vector or environment
}
\keyword{internal}
