% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-vector.R
\name{flatmap}
\alias{flatmap}
\title{Map a function and flatten the result by one-level}
\usage{
flatmap(.x, .f, ..., .type)
}
\description{
Deprecated: please use \code{map()} followed by \code{flatten()}.
}
\keyword{internal}

