% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip the order of the first two arguments of a function}
\usage{
flip(.f)
}
\arguments{
\item{.f}{a function.}
}
\value{
a function where the first two inputs of the input function are
transposed
}
\description{
Flipping the arguments of a function allows you to use a function where the
second argument is the more natural input to use with the pipe (\%>\%)
operator.
}
\examples{
# by default, the rep function repeats each item from the first argument
# n times (from the second argument)
rep(1:3, 3:1)

# using the flip function we instead get n repetitions of the second argument
flip(rep)(1:3, 3:1)

# we can use the flip function with the imap function to create an
# incrementing index where each item is repeated n times
imap(c(2, 1, 4), flip(rep)) \%>\% flatten_int()
}
