% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{has_element}
\alias{has_element}
\title{Does a list contain an object?}
\usage{
has_element(.x, .y)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.y}{Object to test for}
}
\description{
Does a list contain an object?
}
\examples{
x <- list(1:10, 5, 9.9)
x |> has_element(1:10)
x |> has_element(3)
}
