% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{in_new_env}
\alias{in_new_env}
\title{Return a function that evaluates in a new environment.}
\usage{
in_new_env(.f)
}
\arguments{
\item{.f}{The function to evaluate.}
}
\value{
A function that evaluates in a new environment with every call.
}
\description{
Returns a function that creates an environment,
exports its parameters to it, and evaluates the function in the new environment.
The parent is .GlobalEnv.
}
\details{
When calling several functions outside of .GlobalEnv, the entire calling
environment is retained with the object. This can lead to unnecessary information being
retained in memory when moving from a parallel cluster or
saved to disk. Calling a function in an environment
with only the objects required prevents this from happening.
}
\examples{
in_new_env(lm)(Sepal.Length ~ Sepal.Width, data = iris)
}

