% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/objects.R
\name{is_atomic}
\alias{is_atomic}
\title{Is an object atomic}
\usage{
is_atomic(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is_atomic} is equivalent to \code{is.atomic} but does not
return \code{TRUE} for \code{NULL}.
}
\seealso{
bare-predicates
}

