% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/objects.R
\name{is_empty}
\alias{is_empty}
\title{Is a vector/list empty?}
\usage{
is_empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is a vector/list empty?
}
\examples{
is_empty(NULL)
is_empty(list())
is_empty(list(NULL))
}

