% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objects.R
\name{is_index}
\alias{is_index}
\title{Can an object be used as index?}
\usage{
is_index(x)
}
\arguments{
\item{x}{object to test.}
}
\description{
\code{is_index()} checks that \code{x} is either a character or a
numeric vector. It is shorter than directly calling the two type
predicates. It also makes your code clearer by signalling that you
need \code{x} to be an index for an atomic vector or a list.
}

