% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_numeric}
\alias{is_numeric}
\alias{is_scalar_numeric}
\title{Test is an object is integer or double}
\usage{
is_numeric(x)

is_scalar_numeric(x)
}
\description{
Numeric is used in three different ways in base R:
\itemize{
\item as an alias for double (as in \code{\link[=as.numeric]{as.numeric()}})
\item to mean either integer or double (as in \code{\link[=mode]{mode()}})
\item for something representable as numeric (as in \code{\link[=as.numeric]{as.numeric()}})
This function tests for the second, which is often not what you want
so these functions are deprecated.
}
}
\keyword{internal}
