% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objects.R
\name{is_true}
\alias{is_false}
\alias{is_true}
\title{Is an object a logical vector of length 1?}
\usage{
is_true(x)

is_false(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is an object a logical vector of length 1?
}
\examples{
if ("TRUE") is_true("TRUE")
is_false(0)
is_false(FALSE)
}

