% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-flatten.R
\name{list_flatten}
\alias{list_flatten}
\title{Flatten a list}
\usage{
list_flatten(x)
}
\arguments{
\item{x}{A list}
}
\value{
A list, probably longer.
}
\description{
Flattening a list removes a single layer of internal hierarchy.
}
\examples{
x <- list(1, list(2, 3), list(4, list(5)))
x \%>\% list_flatten() \%>\% str()
x \%>\% list_flatten() \%>\% list_flatten() \%>\% str()

# It's now as flat as it can get so further flattening leaves it
# changed.
x \%>\% list_flatten() \%>\% list_flatten() \%>\% list_flatten() \%>\% str()

# Another way to see this is that it reduces the depth of the list
x <- list(
  list(),
  list(list())
)
x \%>\% pluck_depth()
x \%>\% list_flatten() \%>\% pluck_depth()
}
