% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-simplify.R
\name{list_simplify}
\alias{list_simplify}
\title{Simplify a list to an atomic or S3 vector}
\usage{
list_simplify(x, strict = TRUE, ptype = NULL)
}
\arguments{
\item{x}{A list.}

\item{strict}{What should happen if simplification fails? If \code{TRUE},
will error. If \code{FALSE}, will return \code{x} unchanged.}

\item{ptype}{An optional prototype to ensure that the output type is always
the same.}
}
\value{
A vector the same length as \code{x}.
}
\description{
Simplify a list to an atomic or S3 vector
}
\details{
Simplification maintains a one-to-one correspondence between the input
and output, implying that each element of \code{x} must contain a vector of
length 1. If you don't want to maintain this correspondence, then you
probably want either \code{\link[=list_c]{list_c()}} or \code{\link[=list_flatten]{list_flatten()}}.
}
\examples{
list_simplify(list(1, 2, 3))

try(list_simplify(list(1, 2, "x")))
try(list_simplify(list(1, 2, 1:3)))
}
