% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-split.R
\name{list_split}
\alias{list_split}
\title{Split a list or vector into chunks}
\usage{
list_split(x, ..., n = NULL, groups = NULL, is_node = NULL)
}
\arguments{
\item{x}{A list or vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{n}{An integer specifying the size of each chunk. If the length
of \code{x} is not evenly divisible by \code{n}, the last chunk will be smaller.}

\item{groups}{A vector the same length as \code{x} that defines how to
group elements. Elements with the same group value will be placed
in the same chunk.}

\item{is_node}{A predicate function that determines whether an object
is a "node" (by default, a list). See \code{\link[=modify_tree]{modify_tree()}} for more details.}
}
\value{
A list where each element contains a chunk of the original input.
}
\description{
\code{list_split()} splits a list or vector into smaller chunks, providing
a conceptual inverse to \code{\link[=list_flatten]{list_flatten()}}. You can split by chunk size
or by providing a grouping vector.
}
\examples{
# Split by chunk size
list_split(1:10, n = 3)

# Split by grouping vector
list_split(letters[1:6], groups = c(1, 1, 2, 2, 3, 3))

# Split a list
x <- list(a = 1, b = 2, c = 3, d = 4)
list_split(x, n = 2)
}
