% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-modify.R
\name{list_update}
\alias{list_update}
\alias{list_modify}
\alias{list_merge}
\title{Modify a list}
\usage{
list_update(.x, ...)

list_modify(.x, ...)

list_merge(.x, ...)
}
\arguments{
\item{.x}{List to modify.}

\item{...}{New values of a list. Use \code{zap()} to remove values.

These values should be either all named or all unnamed. When
inputs are all named, they are matched to \code{.x} by name. When they
are all unnamed, they are matched by position.

\link[rlang:dyn-dots]{Dynamic dots} are supported. In particular, if your
replacement values are stored in a list, you can splice that in with
\verb{!!!}.}
}
\description{
\itemize{
\item \code{list_update()} modifies the elements of a list by name or position.
\item \code{list_modify()} modifies the elements of a list recursively.
\item \code{list_modify()} merges the elements of a list recursively.
}

\code{list_modify()} is inspired by \code{\link[utils:modifyList]{utils::modifyList()}}.
}
\examples{
x <- list(x = 1:10, y = 4, z = list(a = 1, b = 2))
str(x)

# Update values
str(list_update(x, a = 1))
# Replace values
str(list_update(x, z = 5))
str(list_update(x, z = NULL))

str(list_update(x, z = list(a = 1:5)))
# replace recursively, leaving the other elements of z alone
str(list_modify(x, z = list(a = 1:5)))

# Remove values
str(list_update(x, z = zap()))

# Combine values with list_merge()
str(list_merge(x, x = 11, z = list(a = 2:5, c = 3)))

# All these functions support dynamic dots features. Use !!! to splice
# a list of arguments:
l <- list(new = 1, y = zap(), z = 5)
str(list_update(x, !!!l))
}
