% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-raw.R
\name{map_raw}
\alias{map_raw}
\alias{map2_raw}
\alias{imap_raw}
\alias{pmap_raw}
\alias{flatten_raw}
\title{Functions that return raw vectors}
\usage{
map_raw(.x, .f, ...)

map2_raw(.x, .y, .f, ...)

imap_raw(.x, .f, ...)

pmap_raw(.l, .f, ...)

flatten_raw(.x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions were deprecated in purrr 1.0.0 because they are of limited
use and you can now use \code{map_vec()} instead. They are variants of \code{\link[=map]{map()}},
\code{\link[=map2]{map2()}}, \code{\link[=imap]{imap()}}, \code{\link[=pmap]{pmap()}}, and \code{\link[=flatten]{flatten()}} that return raw vectors.
}
\keyword{internal}
