% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck.R
\name{modify_in}
\alias{modify_in}
\title{Modify a pluck location}
\usage{
modify_in(.x, .where, .f, ...)
}
\arguments{
\item{.x}{A vector or environment}

\item{.where}{A pluck location, as a numeric vector of
positions, a character vector of names, or a list combining both.
The location must exist in the data structure.}

\item{.f}{A function to apply at the pluck location given by \code{.where}.}

\item{...}{Arguments passed to \code{.f}.}
}
\description{
\code{modify_in()} applies a function to a \code{\link[=pluck]{pluck()}} location and
returns the modified data structure.
}
\examples{
# While pluck() returns a component of a data structure that might
# be arbitrarily deep
x <- list(list(bar = 1, foo = 2))
pluck(x, 1, "foo")

# modify_in() applies a function to that location and update the
# element in place:
modify_in(x, list(1, "foo"), ~ .x * 200)

# Additional arguments are passed to the function in the ordinary way:
modify_in(x, list(1, "foo"), `+`, 100)
}
\seealso{
\code{\link[=pluck]{pluck()}}
}
