% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_list.R
\name{modify_list}
\alias{modify_list}
\alias{update_list}
\title{Modify a list}
\usage{
modify_list(x, y)

update_list(`_x`, ...)
}
\arguments{
\item{y}{Replacement list}

\item{_x, x}{List to modify}

\item{...}{New values of a list. Use \code{NULL} to remove values.
Use a formula to evaluate in the context of the list values.}
}
\description{
\code{modify_list()} recursively modifies a list, matching elements either by
name or position. \code{update_list()} is a helper optimised for interactively
modifying an existing list: you can use a formula (a quosure) to replace
using existing values.
}
\examples{
x <- list(x = 1:10, y = 4)
update_list(x, z = 10)
update_list(x, z = ~ x + y)
}
