% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/orify.R
\name{orify}
\alias{orify}
\title{Predicate function operator that creates new predicate functions linked by the || operator.}
\usage{
orify(...)
}
\arguments{
\item{...}{n functions to apply in order from left to right}
}
\value{
A predicate function linked by the || operator.
}
\description{
Predicate function operator that creates new predicate functions linked by the || operator.
}
\examples{
# Examples
is_character_or_factor <- orify(is.character, is.factor)
is_character_or_factor(letters) # TRUE
is_character_or_factor(factor(state.abb)) # TRUE
is_character_or_factor(1:100) # FALSE
}
\seealso{
\code{\link{andify}} to create new predicate functions linked by the && operator.

Other predicate.function.operators: \code{\link{andify}}
}

