% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\docType{package}
\name{purrr-package}
\alias{purrr}
\alias{purrr-package}
\title{purrr: Functional Programming Tools}
\description{
Make your pure functions purr with the 'purrr' package. This
package completes R's functional programming tools with missing features
present in other programming languages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/purrr}
  \item Report bugs at \url{https://github.com/hadley/purrr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item Lionel Henry \email{lionel.hry@gmail.com} [contributor]
  \item RStudio [copyright holder]
}

}
\keyword{internal}
