% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{rate_backoff}
\alias{rate_backoff}
\alias{is_rate}
\title{Create delaying rate settings}
\usage{
rate_backoff(pause_base = 1, pause_cap = 60, pause_min = 1,
  max_times = 3, jitter = TRUE)

is_rate(x)
}
\arguments{
\item{pause_base, pause_cap}{\code{rate_backoff()} uses an exponential
back-off so that each request waits \code{pause_base * 2^i} seconds,
up to a maximum of \code{pause_cap} seconds.}

\item{pause_min}{Minimum time to wait in the backoff; generally
only necessary if you need pauses less than one second (which may
not be kind to the server, use with caution!).}

\item{max_times}{Maximum number of requests to attempt.}

\item{jitter}{Whether to introduce a random jitter in the waiting time.}

\item{x}{An object to test.}
}
\description{
These helpers create rate settings that you can pass to
\code{\link[=insistently]{insistently()}}. You can use them in your own functions with
\code{\link[=rate_sleep]{rate_sleep()}}.
}
\examples{
# With a lower backoff base rate, the waiting time between requests
# will be shorter:
rate_backoff(pause_base = 0.5)

# Pass a minimum waiting time of 5 seconds:
rate_backoff(pause_min = 5)
}
\seealso{
\code{\link[=rate_sleep]{rate_sleep()}}, \code{\link[=insistently]{insistently()}}
}
