% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rdunif}
\alias{rdunif}
\title{Generate random samples from a discrete uniform distribution}
\usage{
rdunif(n, b, a = 1)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{a,b}{Range of the distribution (inclusive).}
}
\description{
Generate random samples from a discrete uniform distribution
}
\examples{
table(rdunif(1e3, 10))
}

