% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-rlang.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_names}
\alias{exec}
\alias{zap}
\alias{\%||\%}
\alias{done}
\alias{rep_along}
\alias{is_bare_list}
\alias{is_bare_atomic}
\alias{is_bare_vector}
\alias{is_bare_double}
\alias{is_bare_integer}
\alias{is_bare_numeric}
\alias{is_bare_character}
\alias{is_bare_logical}
\alias{is_list}
\alias{is_atomic}
\alias{is_vector}
\alias{is_integer}
\alias{is_double}
\alias{is_character}
\alias{is_logical}
\alias{is_null}
\alias{is_function}
\alias{is_scalar_list}
\alias{is_scalar_atomic}
\alias{is_scalar_vector}
\alias{is_scalar_double}
\alias{is_scalar_character}
\alias{is_scalar_logical}
\alias{is_scalar_integer}
\alias{is_empty}
\alias{is_formula}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:\%||\%]{\%||\%}}, \code{\link[rlang:done]{done()}}, \code{\link[rlang:exec]{exec()}}, \code{\link[rlang:is_atomic]{is_atomic()}}, \code{\link[rlang:is_bare_atomic]{is_bare_atomic()}}, \code{\link[rlang:is_bare_character]{is_bare_character()}}, \code{\link[rlang:is_bare_double]{is_bare_double()}}, \code{\link[rlang:is_bare_integer]{is_bare_integer()}}, \code{\link[rlang:is_bare_list]{is_bare_list()}}, \code{\link[rlang:is_bare_logical]{is_bare_logical()}}, \code{\link[rlang:is_bare_numeric]{is_bare_numeric()}}, \code{\link[rlang:is_bare_vector]{is_bare_vector()}}, \code{\link[rlang:is_character]{is_character()}}, \code{\link[rlang:is_double]{is_double()}}, \code{\link[rlang:is_empty]{is_empty()}}, \code{\link[rlang:is_formula]{is_formula()}}, \code{\link[rlang:is_function]{is_function()}}, \code{\link[rlang:is_integer]{is_integer()}}, \code{\link[rlang:is_list]{is_list()}}, \code{\link[rlang:is_logical]{is_logical()}}, \code{\link[rlang:is_null]{is_null()}}, \code{\link[rlang:is_scalar_atomic]{is_scalar_atomic()}}, \code{\link[rlang:is_scalar_character]{is_scalar_character()}}, \code{\link[rlang:is_scalar_double]{is_scalar_double()}}, \code{\link[rlang:is_scalar_integer]{is_scalar_integer()}}, \code{\link[rlang:is_scalar_list]{is_scalar_list()}}, \code{\link[rlang:is_scalar_logical]{is_scalar_logical()}}, \code{\link[rlang:is_scalar_vector]{is_scalar_vector()}}, \code{\link[rlang:is_vector]{is_vector()}}, \code{\link[rlang:rep_along]{rep_along()}}, \code{\link[rlang:set_names]{set_names()}}, \code{\link[rlang:zap]{zap()}}}
}}

