% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-recursive.R
\name{rmodify}
\alias{rmodify}
\title{Recursively modify a list}
\usage{
rmodify(
  x,
  f_leaf = identity,
  f_pre = identity,
  f_post = identity,
  p_leaf = NULL
)
}
\arguments{
\item{x}{A list.}

\item{f_leaf}{A function applied to each leaf.}

\item{f_pre}{A function applied to each node or leaf, before leaves are
transformed.}

\item{f_post}{A function applied to each node or leaf, after leaves are
transformed.}

\item{p_leaf}{A function that determines whether an element is a leaf or a
node. The default value, \code{NULL}, will treats lists as nodes and everything
else as leaves.}
}
\description{
\code{rmodify()} allows you to recursively modify a vector, supplying functions
to either modify each leaf or each node (or both).
}
\examples{
x <- list(list(1, list(2), list(3)))
x \%>\% str()
x \%>\% rmodify(\(x) x + 1) \%>\% str()
}
