% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{scalar-type-predicates}
\alias{scalar-type-predicates}
\alias{is_scalar_list}
\alias{is_scalar_atomic}
\alias{is_scalar_vector}
\alias{is_scalar_numeric}
\alias{is_scalar_integer}
\alias{is_scalar_double}
\alias{is_scalar_character}
\alias{is_scalar_logical}
\title{Scalar type predicates}
\usage{
is_scalar_list(x)

is_scalar_atomic(x)

is_scalar_vector(x)

is_scalar_numeric(x)

is_scalar_integer(x)

is_scalar_double(x)

is_scalar_character(x)

is_scalar_logical(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
These predicates check for a given type and whether the vector is
"scalar", that is, of length 1.
}
\seealso{
\link{type-predicates} \link{bare-type-predicates}
}
