% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{sequences}
\alias{seq_pairs}
\alias{seq_window}
\alias{sequences}
\title{Generate sequences along an object}
\usage{
seq_window(x, size = 2)

seq_pairs(x, filter = `>=`)
}
\arguments{
\item{x}{An object whose length will be used to generate a
sequence.}

\item{size}{The size of the sliding window. Defaults to 2.}

\item{filter}{A predicate function to filter out unwanted
combinations of indices.}
}
\description{
\code{seq_window()} and \code{seq_pairs()} are sequence generators
much like \code{seq_along()} except that they return a list of
indices. They take an object \code{x} and generate a sequence of
indices based on its length.
}
\details{
\code{seq_window()} generates a sliding sequence with a window of
length \code{size} while \code{seq_pairs()} creates a list of
pairwise combinations.
}

