% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_R.R
\name{slide_dbl_R}
\alias{slide_dbl_R}
\title{R implementation of rolling window calculation}
\usage{
slide_dbl_R(.x, .f, .n, .align = "left", .fill = NA, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{.n}{A positive integer; the length of the sliding window. Needs to be
an odd number for centered alignment.}

\item{.align}{One of \code{"left"} (the default), \code{"center"} or \code{"right"};
Assign the result of function calls to the position at the left, center
or right of the sliding window. Ignored if \code{fill = NULL}.}

\item{.fill}{length-one vector or \code{NULL}; value assigned to the result where
the sliding window exceeds the first or last element of \code{.x}. \code{NULL} means
no filling.}

\item{...}{Additional arguments passed on to the mapped function.}
}
\description{
R implementation of \code{\link{slide_dbl}} for speed comparison during
development of the \code{slide} functions.
}
\examples{
x <- rnorm(n = 40)
x \%>\% slide_dbl_R(mean, 20)

# Or use an anonymous function
x \%>\% slide_dbl_R(function(x) mean(x[x > median(x)]), 20)

# Or a formula
x \%>\% slide_dbl_R(~ mean(.x[.x > median(x)]), 20)
}
