% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smash.R
\name{smash}
\alias{smash}
\title{`Smash' a function open to make it take a list of arguments.}
\usage{
smash(.f, ...)
}
\arguments{
\item{.f}{Function to smash.}

\item{...}{Default arguments for \code{.f}. These will be evaluated
once when \code{smash()} is called.}
}
\value{
A function.
}
\description{
Wraps a function in \code{\link{do.call}()}, so instead of taking
multiple arguments, it takes a single named list which will be
interpreted as its arguments. This is useful when you want to pass
a function a row of data frame or list, and don't want to manually
pull it apart in your function. You can also specify default
arguments for \code{.f} in the call to \code{smash()}.
}
\examples{
x <- list(x = c(1:100, NA, 1000), na.rm = TRUE, trim = 0.9)
smash(mean)(x)

# Or in a pipe:
mean \%>\% smash() \%>\% invoke(x)

# Default arguments can also be specified directly in smash()
list(c(1:100, NA, 1000)) \%>\% smash(mean, na.rm = TRUE)()
}
\seealso{
\code{\link{map_call}()} and \code{\link{do.call}()}
}

