% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by.R
\name{split_by}
\alias{split_by}
\title{split_by}
\usage{
split_by(data, ..., sep = ".")
}
\arguments{
\item{data}{a dataframe or tibble}

\item{...}{bare symbols for columns names in \code{data} to use for splitting.
Passed on to \code{dplyr::group_by()}}

\item{sep}{If splitting with multiple variables, sep is used to create names.}
}
\value{
a named list of dataframes
}
\description{
An alternative to \code{\link[base:split.data.frame]{base::split.data.frame()}} with the following differences:
\itemize{
\item it operates on bare symbols (like dplyr verbs)
\item if splitting on multiple variables, it does not include missing combinations
not present in the data (equivelatnt to \code{drop = TRUE} in
\code{\link[base:split.data.frame]{base::split.data.frame()}})
}
}
\details{
This funciton is powerful in combination with \code{\link[=map_dfr]{map_dfr()}}.
}
\examples{
split_by(mtcars, cyl)
split_by(mtcars, cyl, gear)
if(require(dplyr))
  split_by(mtcars, row_number())

if (require(ggplot2)) {
  split_by(mpg, manufacturer)
  split_by(mpg, manufacturer, class)
}
}
