% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{zip}
\alias{unzip}
\alias{zip}
\title{Zip and unzip lists.}
\usage{
zip(.x)

unzip(.x, fields = NULL)
}
\arguments{
\item{.x}{A list.}

\item{fields}{Fields to use when unzipping - defaults to the names
of the first sub-list.}
}
\description{
Zip and unzip are useful transforms for turning a list-of-lists "inside-out".
For example, if you had a list of length n where each component had values
\code{a} and \code{b}, \code{unzip} would make a with elements \code{a}
and \code{b} that contained lists of length n.
}
\examples{
x <- rerun(10, x = runif(1), y = runif(5))
unzip(x)

zip(list(a = 1:5, b = 5:1))
}

