% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_n}
\alias{zip2}
\alias{zip3}
\alias{zip_n}
\title{Transpose a list.}
\usage{
zip_n(.l, .fields = NULL, .simplify = FALSE)

zip2(.x, .y, .fields = NULL, .simplify = FALSE)

zip3(.x, .y, .z, .fields = NULL, .simplify = FALSE)
}
\arguments{
\item{.l}{A list of lists or vectors to zip.}

\item{.fields}{Fields to use when unzipping - defaults to the names
of the first sub-list.}

\item{.simplify}{If \code{TRUE}, lists containing atomic scalars of
the same type will be converted to a vector.}

\item{.x, .y, .z}{Lists or vectors to zip.}
}
\description{
Zip turns a list-of-lists "inside-out"; it turns a pair of lists into a
list of pairs, or a list of pairs into pair of lists. For example,
If you had a list of length n where each component had values \code{a} and
\code{b}, \code{zip_n()} would make a list with elements \code{a} and
\code{b} that contained lists of length n.
}
\details{
Note that \code{zip_n()} is its own inverse, much like the
transpose operation on a matrix. You can get back the original
input by zipping it twice. \code{zip2(x, y)} is a shortcut for
\code{zip_n(list(x, y))}.
}
\examples{
x <- rerun(5, x = runif(1), y = runif(5))
x \%>\% str()
x \%>\% zip_n() \%>\% str()
# Back to where we started
x \%>\% zip_n() \%>\% zip_n() \%>\% str()

# zip() is useful in conjunction with safe()
x <- list("a", 1, 2)
y <- x \%>\% map(safe(log))
y \%>\% str()
y \%>\% zip_n() \%>\% str()

# The simplify argument reduces list to atomic vectors where possible
x <- list(a = 1:5, b = 5:1)
x \%>\% zip_n()
x \%>\% zip_n(.simplify = TRUE)
}

