% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/along.R
\name{along}
\alias{along}
\alias{list_along}
\title{Create a list of given length}
\usage{
list_along(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A list of the same length as \code{x}.
}
\description{
\Sexpr[results=rd, stage=render]{purrr:::lifecycle("questioning")}

It can be useful to create an empty list that you plan to fill later. This is
similar to the idea of \code{\link[=seq_along]{seq_along()}}, which creates a vector of the same
length as its input.
}
\details{
This function might change to \code{\link[vctrs:vec_init]{vctrs::vec_init()}}.
}
\examples{
x <- 1:5
seq_along(x)
list_along(x)
}
\keyword{internal}
