% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-map.R
\name{at_depth}
\alias{at_depth}
\title{Map at depth}
\usage{
at_depth(.x, .depth, .f, ...)
}
\description{
This function is defunct and has been replaced by \code{\link[=map_depth]{map_depth()}}.
See also \code{\link[=modify_depth]{modify_depth()}} for a version that preserves the types of
the elements of the tree.
}
\keyword{internal}
