% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{compose}
\alias{compose}
\title{Compose multiple functions}
\usage{
compose(...)
}
\arguments{
\item{...}{n functions to apply in order from right to left.}
}
\description{
Compose multiple functions
}
\examples{
not_null <- compose(`!`, is.null)
not_null(4)
not_null(NULL)

add1 <- function(x) x + 1
compose(add1, add1)(8)
}

