% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objects.R
\name{enlist}
\alias{enlist}
\title{Concatenate objects in a list}
\usage{
enlist(...)
}
\arguments{
\item{...}{Objects to concatenate.}
}
\description{
This joins all arguments in a list. Non-list objects are
encapsulated in a list before concatenation.
}
\examples{
inputs <- list(arg1 = "a", arg2 = "b")

# enlist() concatenates the elements of inputs with arg3
enlist(inputs, arg3 = c("c1", "c2")) \%>\% str()
list(inputs, arg3 = c("c1", "c2")) \%>\% str()
c(inputs, arg3 = c("c1", "c2")) \%>\% str()
}

