% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every-some.R
\name{every}
\alias{every}
\alias{some}
\title{Do every or some elements of a list satisfy a predicate?}
\usage{
every(.x, .p, ...)

some(.x, .p, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.p}{A predicate function to apply on each element of \code{.x}.
\code{some()} returns \code{TRUE} when \code{.p} is \code{TRUE} for at least one
element. \code{every()} returns \code{TRUE} when \code{.p} is \code{TRUE} for all
elements.}

\item{...}{Additional arguments passed on to \code{.p}.}
}
\value{
A logical vector of length 1.
}
\description{
Do every or some elements of a list satisfy a predicate?
}
\examples{
x <- list(0, 1, TRUE)
x \%>\% every(identity)
x \%>\% some(identity)

y <- list(0:10, 5.5)
y \%>\% every(is.numeric)
y \%>\% every(is.integer)
}
