% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{every}
\alias{every}
\alias{some}
\title{Do every or some elements of a list satisfy a predicate?}
\usage{
every(.x, .f, ...)

some(.x, .f, ...)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A predicate function, i.e. a function that returns either
\code{TRUE} or \code{FALSE}}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Do every or some elements of a list satisfy a predicate?
}
\examples{
x <- list(0, 1, TRUE)
x \%>\% every(identity)
x \%>\% some(identity)

y <- list(0:10, 5.5)
y \%>\% every(is.numeric)
y \%>\% every(is.integer)
}

