% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Flatten a list of lists into a list.}
\usage{
flatten(x, recursive = FALSE)
}
\arguments{
\item{x}{A list of flatten}

\item{recursive}{Recursively flatten list components?}
}
\description{
This is a wrapper for \code{\link{unlist}()} that defaults to
\code{recursive = FALSE}.
}
\examples{
x <- rerun(10, sample(5))
x \%>\% flatten(x)

# You can use flatten in conjunction with map
x \%>\% map(~ .[1]) \%>\% flatten()
# But it's better to use map_v
x \%>\% map_v(~ .[1])
}

