% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get-attr}
\alias{get-attr}
\alias{\%@\%}
\title{Infix attribute accessor}
\usage{
x \%@\% name
}
\arguments{
\item{x}{Object}

\item{name}{Attribute name}
}
\description{
\Sexpr[results=rd, stage=render]{purrr:::lifecycle("soft-deprecated")}

Please use the \verb{\%@\%} operator exported in rlang. It has an
interface more consistent with \code{@}: uses NSE, supports S4 fields,
and has an assignment variant.
}
\examples{
factor(1:3) \%@\% "levels"
mtcars \%@\% "class"
}
\keyword{internal}
