% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-utils.R
\name{get-attr}
\alias{get-attr}
\alias{\%@\%}
\title{Infix attribute accessor}
\usage{
x \%@\% name
}
\arguments{
\item{x}{Object}

\item{name}{Attribute name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in purrr 0.3.0. Instead, lease use the \verb{\%@\%}
operator exported in rlang. It has an interface more consistent with \code{@}:
uses NSE, supports S4 fields, and has an assignment variant.
}
\keyword{internal}
