% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-tail.R
\name{head_while}
\alias{head_while}
\alias{tail_while}
\title{Find head/tail that all satisfies a predicate.}
\usage{
head_while(.x, .p, ...)

tail_while(.x, .p, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.p}{A predicate function (i.e. a function that returns either \code{TRUE}
or \code{FALSE}) specified in one of the following ways:
\itemize{
\item A named function, e.g. \code{is.character}.
\item An anonymous function, e.g. \verb{\\(x) all(x < 0)} or \code{function(x) all(x < 0)}.
\item A formula, e.g. \code{~ all(.x < 0)}. You must use \code{.x} to refer to the first
argument). Only appropriate for very tasks.
}}

\item{...}{Additional arguments passed on to the mapped function.

Note that the arguments that differ in each call come before \code{.f},
and the arguments that are the same come after \code{.f}.}
}
\value{
A vector the same type as \code{.x}.
}
\description{
Find head/tail that all satisfies a predicate.
}
\examples{
pos <- function(x) x >= 0
head_while(5:-5, pos)
tail_while(5:-5, negate(pos))

big <- function(x) x > 100
head_while(0:10, big)
tail_while(0:10, big)
}
