% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invoke}
\alias{invoke}
\alias{invoke_v}
\title{Invoke every element of a list of functions.}
\usage{
invoke(.x, ...)

invoke_v(.x, ..., .type)
}
\arguments{
\item{.x}{A function or list of functions.}

\item{...}{Arguments passed to each function.}

\item{.type}{Specifies the type of result of \code{.f}, if known.
This will improve performance, and adds a test that output of \code{.f}
is the type that you expect.}
}
\description{
Invoke every element of a list of functions.
}
\examples{
middle <- list(m1 = mean, m2 = median)

middle \%>\% invoke(rcauchy(100))
middle \%>\% invoke_v(rcauchy(100))
}

