% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{invoke}
\alias{invoke}
\alias{invoke_chr}
\alias{invoke_dbl}
\alias{invoke_int}
\alias{invoke_lgl}
\title{Invoke every element of a list of functions.}
\usage{
invoke(.x, ...)

invoke_lgl(.x, ...)

invoke_int(.x, ...)

invoke_dbl(.x, ...)

invoke_chr(.x, ...)
}
\arguments{
\item{.x}{A function or list of functions.}

\item{...}{Arguments passed to each function.}
}
\description{
Invoke every element of a list of functions.
}
\examples{
middle <- list(m1 = mean, m2 = median)

middle \%>\% invoke(rcauchy(100))
middle \%>\% invoke_dbl(rcauchy(100))
}

