% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{keep}
\alias{compact}
\alias{discard}
\alias{keep}
\title{Keep or discard elements using a predicate function.}
\usage{
keep(.x, .f, ...)

discard(.x, .f, ...)

compact(.x)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A predicate function, i.e. a function that either returns
\code{TRUE} or \code{FALSE}}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\description{
\code{keep} and \code{discard} are opposites. \code{compact} is a handy
wrapper that removes all elements that are \code{NULL}.
}
\details{
These are usually called \code{select} or \code{filter} and \code{reject} or
\code{drop}, but those names are already taken. \code{keep} is similar to
\code{\link{Filter}} but the argument order is more convenient, and the
evaluation of \code{.f} is stricter.
}
\examples{
rep(10, 10) \%>\%
  map(sample, 5) \%>\%
  keep(function(x) mean(x) > 6)

# You can also create functions with \%>\%
rep(10, 10) \%>\%
  map(sample, 5) \%>\%
  keep(. \%>\% {mean(.) > 6})
}

