% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-combine.R
\name{list_c}
\alias{list_c}
\alias{list_cbind}
\alias{list_rbind}
\title{Combine list elements into a single data structure}
\usage{
list_c(x, ptype = NULL)

list_cbind(
  x,
  name_repair = c("unique", "universal", "check_unique"),
  size = NULL
)

list_rbind(x, id = NULL, ptype = NULL)
}
\arguments{
\item{x}{A list.}

\item{ptype}{An optional prototype to ensure that the output type is always
the same.}

\item{name_repair}{One of \code{"unique"}, \code{"universal"}, or \code{"check_unique"}.
See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for the meaning of these options.}

\item{size}{An optional integer size to ensure that every input has the
same size (i.e. number of rows).}
}
\description{
\itemize{
\item \code{list_c()} combines elements into a vector by concatenating them together
with \code{\link[vctrs:vec_c]{vctrs::vec_c()}}.
\item \code{list_rbind()} combines elements into a data frame by row-binding them
together with \code{\link[vctrs:vec_bind]{vctrs::vec_rbind()}}.
\item \code{list_cbind()} combines elements into a data frame by column-binding them
together with \code{\link[vctrs:vec_bind]{vctrs::vec_cbind()}}.
}
}
\examples{
x1 <- list(a = 1, b = 2, c = 3)
list_c(x1)

x2 <- list(
  data.frame(x = 1:2),
  data.frame(y = "a")
)
list_rbind(x2)
list_cbind(x2)
}
