% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{map}
\alias{each}
\alias{map}
\title{Apply a function to each element of a list.}
\usage{
map(.x, .f, ..., .type)

each(.x, .f, ...)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A function to apply to each element of \code{.x}.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.type}{Specifies the type of result of \code{.f}, if known.
If supplied, the result of \code{map} will be a vector or matrix.
If omitted, the result of \code{map} will be a list.}
}
\value{
\code{map} the transformed input; \code{each} the input \code{.x}.
}
\description{
\code{map} returns the transformed input; \code{each} calls \code{.f} for
its side-effect and returns the original input.
}
\examples{
1:10 \%>\%
  map(rnorm, n = 10) \%>\%
  map(mean, .type = numeric(1))
}
\seealso{
\code{\link{pluck}} for a wrapper around a common use case:
  extracting an element from each component; \code{\link{map2}()} and
  \code{\link{map3}()} to map over multiple inputs simulatenously
}

