% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{map2}
\alias{map2}
\alias{map3}
\title{Map over multiple inputs simultaneously.}
\usage{
map2(.x, .y, .f, ...)

map3(.x, .y, .z, .f, ...)
}
\arguments{
\item{.x,.y,.z}{Lists, usually of the same length. If not, lists will
be recycled to the length of the longest, using R's regular recycling
rules.}

\item{.f}{A function to apply to each element of \code{.x}.}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\description{
Map over multiple inputs simultaneously.
}
\examples{
x <- list(1, 10, 100)
y <- list(1, 2, 3)
map2(x, y, `+`)
}

