% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map.R
\name{map2}
\alias{map2}
\alias{map3}
\title{Map over multiple inputs simultaneously.}
\usage{
map2(.x, .y, .f, ...)

map3(.x, .y, .z, .f, ...)
}
\arguments{
\item{.x,.y,.z}{Lists, usually of the same length. If not, lists will
be recycled to the length of the longest, using R's regular recycling
rules.}

\item{.f}{A function of two (for \code{map2}) or three (\code{map3})
arguments.}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\description{
These functions are designed in such a way that arguments to be vectorised
over come before the function name, and arguments that should be supplied to
every call come after the function name.
}
\examples{
x <- list(1, 10, 100)
y <- list(1, 2, 3)
map2(x, y, `+`)

# Split into pieces, fit model to each piece, then predict
by_cyl <- mtcars \%>\% split(.$cyl)
mods <- by_cyl \%>\% map(~ lm(mpg ~ wt, data = .))
map2(mods, by_cyl, predict)
}

