% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{map_if}
\alias{map_if}
\title{Modify elements where predicate is satisified.}
\usage{
map_if(.x, .p, .f, ...)
}
\arguments{
\item{.x}{A list or vector.}

\item{.p}{A predicate function. Only those elements where \code{.p}
evaluates to \code{TRUE} will be modified.}

\item{.f}{A function, formula or string.

  If a function, it is used as is.

  If a formula, e.g. \code{~ x + 2}, it is converted to a function with
  a single argument, \code{x}, e.g. \code{function(x) x + 2}. This is less
  typing for very simple anonymous functions.

  If a string, e.g. \code{"y"}, it is converted to an extractor function,
  \code{function(x) x[["y"]]}.}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\description{
Modify elements where predicate is satisified.
}
\examples{
x <- rerun(10, y = if (rbinom(1, 1, prob = 0.5) == 1) NULL else sample(100, 5))
x \%>\% map_if(~ !is.null(x$y), ~ update_list(x, y = ~ y * 100))
}

