% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/negate.R
\name{negate}
\alias{negate}
\title{Negate a predicate function.}
\usage{
negate(.f)
}
\arguments{
\item{.f}{A function, formula or string.

  If a function, it is used as is.

  If a formula, e.g. \code{~ x + 2}, it is converted to a function with
  a single argument, \code{x}, e.g. \code{function(x) x + 2}. This is less
  typing for very simple anonymous functions.

  If a string, e.g. \code{"y"}, it is converted to an extractor function,
  \code{function(x) x[["y"]]}.}
}
\value{
A new predicate function.
}
\description{
Negate a predicate function.
}
\examples{
x <- zip(list(x = 1:10, y = rbenoulli(10)))
x \%>\% keep("y") \%>\% length()
x \%>\% keep(negate("y")) \%>\% length()
# Same as
x \%>\% discard("y") \%>\% length()
}

