% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverb-negate.R
\name{negate}
\alias{negate}
\title{Negate a predicate function}
\usage{
negate(.p)
}
\arguments{
\item{.p}{A predicate function or a formula.}
}
\value{
A new predicate function.
}
\description{
Negating a function changes \code{TRUE} to \code{FALSE} and \code{FALSE} to \code{TRUE}.
}
\section{Adverbs}{
This function is called an adverb because it modifies the effect of a
function (a verb). If you'd like to include a function created an adverb
in a package, be sure to read \link{faq-adverbs-export}.
}

\examples{
x <- list(x = 1:10, y = rbernoulli(10), z = letters)
x \%>\% keep(is.numeric) \%>\% names()
x \%>\% keep(negate(is.numeric)) \%>\% names()
# Same as
x \%>\% discard(is.numeric)
}
\seealso{
Other adverbs: 
\code{\link{auto_browse}()},
\code{\link{compose}()},
\code{\link{possibly}()},
\code{\link{quietly}()},
\code{\link{safely}()}
}
\concept{adverbs}
