% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{parallelization}
\alias{parallelization}
\title{Parallelization in purrr}
\description{
purrr's map functions have a \code{.parallel} argument to parallelize a map using
the \CRANpkg{mirai} package. This allows you to run computations in parallel
using more cores on your machine, or distributed over the network.

Parallelizing a map using 'n' processes does not automatically lead to it
taking 1/n of the time. Additional overhead from setting up the parallel task
and communicating with parallel processes eats into this benefit, and can
outweigh it for very short tasks or those involving large amounts of data.
The threshold at which parallelization becomes clearly beneficial will differ
according to your individual setup and task, but a rough guide would be in
the order of 100 microseconds to 1 millisecond for each map iteration.
}
\section{Daemons settings}{
How and where parallelization occurs is determined by
\code{\link[mirai:daemons]{daemons()}}. This is a function from the \pkg{mirai}
package that sets up daemons (persistent background processes that receive
parallel computations) on your local machine or across the network.

\pkg{purrr} requires these to be set up prior to performing any parallel map
operations. It is usual to set daemons once per session. You can leave them
running as they consume almost no resources whilst waiting to receive tasks.
The following sets up 6 daemons on your local machine:

\preformatted{
daemons(6)
}

\code{daemons()}arguments:
\itemize{
\item \code{n}: the number of daemons to launch on your local machine, e.g.
\code{daemons(6)}. As a rule of thumb, for maximum efficiency this should be (at
most) one less than the number of cores on your machine, leaving one core
for the main R process.
\item \code{url} and \code{remote}: used to set up and launch daemons for distributed
computing over the network. See \link[mirai:daemons]{daemons} function
documentation for more details.
\item None: calling \code{daemons()} with no arguments returns a summary of the
current connection status and mirai tasks queued and completed.
}

For more details, including on further arguments, see
\link[mirai:daemons]{daemons}.

Resetting daemons:

Daemons persist for the duration of your session.

\code{daemons(0)} resets and terminates any existing daemons.

All daemons automatically terminate when your session ends and the connection
drops. Hence you do not need to explicitly terminate daemons in this instance,
although it is still good practice to do so.
}

\section{Crating a function}{
As the function \code{.f} needs to be serialized and shared with other processes,
we recommend that you \link[carrier:crate]{crate()} your function for use in a
parallel map.

The only exception is if \code{.f} is a function from a package or base R. Do not
\link[carrier:crate]{crate} these functions - although you can
\link[carrier:crate]{crate} an anonymous function that calls these functions.
\preformatted{
# either use a base R or package function directly:
mtcars |> map_dbl(sum, .parallel = TRUE)
# or crate() it as an anonymous function:
mtcars |> map_dbl(crate(function(...) sum(...)), .parallel = TRUE)
}

Crating ensures that everything needed by the function is serialized along
with it, but not other objects which happen to be in the function's enclosing
environment. This helps to prevent inadvertently shipping large data objects
to daemons, and is hence recommnded practice, even when a function is
otherwise self-contained.

Examples:
\preformatted{
# do not crate() a package function:
map(1:3, stats::runif, .parallel = TRUE)

# crate() a function, including the definition of 'fun':
fun <- function(x) \{x + x \%\% 2 \}
map(1:3, crate(function(x) x + fun(x), fun = fun), .parallel = TRUE)
}

For details on further options, see \link[carrier:crate]{crate}.
}

\section{Further documentation}{
\pkg{purrr}'s parallelization is powered by \CRANpkg{mirai}.See the
\href{https://shikokuchuo.net/mirai/articles/mirai.html}{mirai introduction and reference}
for more details.

\link{crate} is re-exported from the \CRANpkg{carrier} package. See the
\href{https://github.com/r-lib/carrier}{crate readme} for more details.
}

