% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{pluck}
\alias{pluck}
\title{Extract elements of a list by position.}
\usage{
pluck(.x, .i, .type)
}
\arguments{
\item{.x}{A list or vector.}

\item{.i}{A string or integer index.}

\item{.type}{Specifies the type of result of \code{.f}, if known.
If supplied, the result of \code{map} will be a vector or matrix.
If omitted, the result of \code{map} will be a list.}
}
\description{
Extract elements of a list by position.
}
\examples{
by_cyl <- split(mtcars, mtcars$cyl)
mod_sum <- by_cyl \%>\%
  map(lm, formula = mpg ~ wt) \%>\%
  map(summary)

# If .type is missing, you always get a list
mod_sum \%>\% pluck("r.squared")

# If you supply .type, you'll get a vector
mod_sum \%>\% pluck("r.squared", numeric(1))
}

