% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mapper.R
\name{pluck}
\alias{pluck}
\title{Pluck out a single an element from a vector or environment}
\usage{
pluck(x, index, default = NULL)
}
\arguments{
\item{x}{A vector or environment}

\item{index}{A list indexing into the object}

\item{default}{Value to use if target is empty or absent.}
}
\description{
This is a generalise form of \code{[[} which allows you to index by
name, position, or attribute.
}
\keyword{internal}
