% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\docType{package}
\name{purrr-package}
\alias{purrr}
\alias{purrr-package}
\title{purrr: Functional Programming Tools}
\description{
A complete and consistent functional programming toolkit for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://purrr.tidyverse.org}
  \item \url{https://github.com/tidyverse/purrr}
  \item Report bugs at \url{https://github.com/tidyverse/purrr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item Lionel Henry \email{lionel.hry@gmail.com} [contributor]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
