% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\docType{package}
\name{purrr-package}
\alias{purrr}
\alias{purrr-package}
\title{purrr: Functional Programming Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A complete and consistent functional programming
    toolkit for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://purrr.tidyverse.org}
  \item \url{https://github.com/tidyverse/purrr}
  \item Report bugs at \url{https://github.com/tidyverse/purrr/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
