% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-purrr.R
\docType{package}
\name{purrr-package}
\alias{purrr}
\alias{purrr-package}
\title{purrr: Functional Programming Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A complete and consistent functional programming toolkit for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://purrr.tidyverse.org/}
  \item \url{https://github.com/tidyverse/purrr}
  \item Report bugs at \url{https://github.com/tidyverse/purrr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (https://ror.org/03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
