% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-utils.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Generate random sample from a Bernoulli distribution}
\usage{
rbernoulli(n, p = 0.5)
}
\arguments{
\item{n}{Number of samples}

\item{p}{Probability of getting \code{TRUE}}
}
\value{
A logical vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in purrr 1.0.0 because it's not related to the
core purpose of purrr.
}
\examples{
rbernoulli(10)
rbernoulli(100, 0.1)
}
\keyword{internal}
