% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Generate random sample from a Bernoulli distribution}
\usage{
rbernoulli(n, p = 0.5)
}
\arguments{
\item{n}{Number of samples}

\item{p}{Probability of getting \code{TRUE}}
}
\value{
A logical vector
}
\description{
Generate random sample from a Bernoulli distribution
}
\examples{
rbernoulli(10)
rbernoulli(100, 0.1)
}
