% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-utils.R
\name{rdunif}
\alias{rdunif}
\title{Generate random sample from a discrete uniform distribution}
\usage{
rdunif(n, b, a = 1)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{a, b}{Range of the distribution (inclusive).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in purrr 1.0.0 because it's not related to the
core purpose of purrr.
}
\examples{
table(rdunif(1e3, 10))
table(rdunif(1e3, 10, -5))
}
\keyword{internal}
