% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdunif}
\alias{rdunif}
\title{Generate random sample from a discrete uniform distribution}
\usage{
rdunif(n, b, a = 1)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{a, b}{Range of the distribution (inclusive).}
}
\description{
Generate random sample from a discrete uniform distribution
}
\examples{
table(rdunif(1e3, 10))
table(rdunif(1e3, 10, -5))
}
