% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{reduce}
\alias{reduce}
\alias{reduce_right}
\title{Reduce a list to a single value by iteratively applying a binary function.}
\usage{
reduce(.x, .f, ..., .init)

reduce_right(.x, .f, ..., .init)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A two-argument function.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.init}{If supplied, will be used as the first value to start
the accumulation, rather than using \code{x[[1]]}.}
}
\description{
\code{reduce} combines from the left, \code{reduce_right} combines from
the right.
}
\examples{
1:3 \%>\% reduce(`+`)
1:10 \%>\% reduce(`*`)

5 \%>\%
  replicate(sample(10, 5), simplify = FALSE) \%>\%
  reduce(intersect)

x <- list(c(0, 1), c(2, 3), c(4, 5))
x \%>\% reduce(c)
x \%>\% reduce_right(c)
# Equivalent to:
x \%>\% rev() \%>\% reduce(c)
}

