% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce_right}
\alias{reduce_right}
\alias{reduce2_right}
\alias{accumulate_right}
\title{Reduce from the right (retired)}
\usage{
reduce_right(.x, .f, ..., .init)

reduce2_right(.x, .y, .f, ..., .init)

accumulate_right(.x, .f, ..., .init)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{For \code{reduce()}, and \code{accumulate()}, a 2-argument
function. The function will be passed the accumulated value as
the first argument and the "next" value as the second argument.

For \code{reduce2()} and \code{accumulate2()}, a 3-argument function. The
function will be passed the accumulated value as the first
argument, the next value of \code{.x} as the second argument, and the
next value of \code{.y} as the third argument.

The reduction terminates early if \code{.f} returns a value wrapped in
a \code{\link[=done]{done()}}.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.init}{If supplied, will be used as the first value to start
the accumulation, rather than using \code{.x[[1]]}. This is useful if
you want to ensure that \code{reduce} returns a correct value when \code{.x}
is empty. If missing, and \code{.x} is empty, will throw an error.}

\item{.y}{For \code{reduce2()} and \code{accumulate2()}, an additional
argument that is passed to \code{.f}. If \code{init} is not set, \code{.y}
should be 1 element shorter than \code{.x}.}
}
\description{
\Sexpr[results=rd, stage=render]{purrr:::lifecycle("soft-deprecated")}

These functions are retired as of purrr 0.3.0. Please use the
\code{.dir} argument of \code{\link[=reduce]{reduce()}} instead, or reverse your vectors
and use a left reduction.
}
\keyword{internal}
