% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{is_bare_list}
\alias{reexports}
\alias{is_bare_atomic}
\alias{reexports}
\alias{is_bare_vector}
\alias{reexports}
\alias{is_bare_double}
\alias{reexports}
\alias{is_bare_integer}
\alias{reexports}
\alias{is_bare_numeric}
\alias{reexports}
\alias{is_bare_character}
\alias{reexports}
\alias{is_bare_logical}
\alias{reexports}
\alias{is_list}
\alias{reexports}
\alias{is_atomic}
\alias{reexports}
\alias{is_vector}
\alias{reexports}
\alias{is_integer}
\alias{reexports}
\alias{is_double}
\alias{reexports}
\alias{is_character}
\alias{reexports}
\alias{is_logical}
\alias{reexports}
\alias{is_null}
\alias{reexports}
\alias{is_function}
\alias{reexports}
\alias{is_scalar_list}
\alias{reexports}
\alias{is_scalar_atomic}
\alias{reexports}
\alias{is_scalar_vector}
\alias{reexports}
\alias{is_scalar_double}
\alias{reexports}
\alias{is_scalar_character}
\alias{reexports}
\alias{is_scalar_logical}
\alias{reexports}
\alias{is_scalar_integer}
\alias{reexports}
\alias{is_empty}
\alias{reexports}
\alias{is_formula}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang]{is_bare_list}}, \code{\link[rlang]{is_bare_atomic}}, \code{\link[rlang]{is_bare_vector}}, \code{\link[rlang]{is_bare_double}}, \code{\link[rlang]{is_bare_integer}}, \code{\link[rlang]{is_bare_numeric}}, \code{\link[rlang]{is_bare_character}}, \code{\link[rlang]{is_bare_logical}}, \code{\link[rlang]{is_list}}, \code{\link[rlang]{is_atomic}}, \code{\link[rlang]{is_vector}}, \code{\link[rlang]{is_integer}}, \code{\link[rlang]{is_double}}, \code{\link[rlang]{is_character}}, \code{\link[rlang]{is_logical}}, \code{\link[rlang]{is_null}}, \code{\link[rlang]{is_function}}, \code{\link[rlang]{is_scalar_list}}, \code{\link[rlang]{is_scalar_atomic}}, \code{\link[rlang]{is_scalar_vector}}, \code{\link[rlang]{is_scalar_double}}, \code{\link[rlang]{is_scalar_character}}, \code{\link[rlang]{is_scalar_logical}}, \code{\link[rlang]{is_scalar_integer}}, \code{\link[rlang]{is_empty}}, \code{\link[rlang]{is_formula}}}
}}

