% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names}
\alias{set_names}
\title{Set names in a vector}
\usage{
set_names(x, nm = x)
}
\arguments{
\item{x}{Vector to name}

\item{nm}{Vector of names, the same length as \code{x}}
}
\description{
This is a snake case wrapper for \code{\link[stats]{setNames}}, with
tweaked defaults, and stricter argument checking.
}
\examples{
set_names(1:4, c("a", "b", "c", "d"))

# If the second argument is ommitted a vector is named with itself
set_names(letters[1:5])
}

