% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objects.R
\name{type-predicates}
\alias{is_atomic}
\alias{is_character}
\alias{is_integer}
\alias{is_list}
\alias{is_logical}
\alias{is_numeric}
\alias{is_vector}
\alias{type-predicates}
\title{Type predicates}
\usage{
is_list()

is_atomic(x)

is_vector(x)

is_numeric()

is_integer()

is_character()

is_logical()
}
\arguments{
\item{x}{object to be tested.}
}
\description{
These type predicates aim to make type testing in R more
consistent.
}
\details{
Most of these predicates are simple aliases to base R functions. In
addition:
\itemize{
  \item Unlike \code{is.atomic()}, \code{is_atomic()} does not
        return \code{TRUE} for \code{NULL}.
  \item Unlike \code{is.vector()}, \code{is_vector()} test if an
        object is an atomic vector or a list. \code{is.vector}
        checks for the presence of attributes (other than name).
  \item Unlike \code{is.numeric()}, \code{is_numeric()} only
        returns \code{TRUE} for floating point numbers, not
        integers.
}
}
\seealso{
bare-type-predicates
}

