% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Modify a list}
\usage{
update_list(`_data`, ...)
}
\arguments{
\item{_data}{A list.}

\item{...}{New values of a list. Use \code{NULL} to remove values.
Use a formula to evaluate in the context of the list values.}
}
\description{
Modify a list
}
\examples{
x <- list(x = 1:10, y = 4)
update_list(x, z = 10)
update_list(x, z = ~ x + y)
}
