% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-modify.R
\name{update_list}
\alias{update_list}
\title{Update a list with formulas}
\usage{
update_list(.x, ...)
}
\arguments{
\item{.x}{List to modify.}

\item{...}{New values of a list. Use \code{zap()} to remove values.

These values should be either all named or all unnamed. When
inputs are all named, they are matched to \code{.x} by name. When they
are all unnamed, they are matched by position.

\link[rlang:dyn-dots]{Dynamic dots} are supported. In particular, if your
replacement values are stored in a list, you can splice that in with
\verb{!!!}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{update_list()} was deprecated in purrr 1.0.0, because we no longer believe
that functions that use NSE are a good fit for purrr.

\code{update_list()} handles formulas and quosures that can refer to
values existing within the input list. This function is deprecated
because we no longer believe that functions that use tidy evaluation are
a good fit for purrr.
}
\keyword{internal}
