% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip.R
\name{zip}
\alias{unzip}
\alias{zip}
\title{Zip and unzip lists.}
\usage{
zip(.x)

unzip(.x, .fields = NULL, .simplify = TRUE)
}
\arguments{
\item{.x}{A list.}

\item{fields}{Fields to use when unzipping - defaults to the names
of the first sub-list.}

\item{simplify}{If \code{TRUE}, lists containing atomic scalars of the
same type will be converted to a vector.}
}
\description{
Zip and unzip are useful transforms for turning a list-of-lists "inside-out".
For example, if you had a list of length n where each component had values
\code{a} and \code{b}, \code{unzip} would make a with elements \code{a}
and \code{b} that contained lists of length n.
}
\examples{
x <- rerun(5, x = runif(1), y = runif(5))
str(x)
unzip(x)
unzip(x, simplify = TRUE)

zip(list(a = 1:5, b = 5:1))
# unzip is similar to map2 used with list()
map2(1:5, 5:1, list)
}

